Move tree algorithms to its own file
