+++ src/tree.rs
@@ -1,0 +1,140 @@
+// Licensed under the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License. You may obtain
+// a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+// License for the specific language governing permissions and limitations
+// under the License.
+
+#![allow(clippy::float_cmp)]
+
+use std::cmp::Ordering;
+
+use super::{graph, weight_callable};
+
+use pyo3::exceptions::PyValueError;
+use pyo3::prelude::*;
+use pyo3::Python;
+
+use petgraph::prelude::*;
+use petgraph::stable_graph::EdgeReference;
+use petgraph::unionfind::UnionFind;
+use petgraph::visit::{IntoEdgeReferences, NodeIndexable};
+
+use rayon::prelude::*;
+
+use crate::iterators::WeightedEdgeList;
+
+/// Find the edges in the minimum spanning tree or forest of a graph
+/// using Kruskal's algorithm.
+///
+/// :param PyGraph graph: Undirected graph
+/// :param weight_fn: A callable object (function, lambda, etc) which
+///     will be passed the edge object and expected to return a ``float``. This
+///     tells retworkx/rust how to extract a numerical weight as a ``float``
+///     for edge object. Some simple examples are::
+///
+///         minimum_spanning_edges(graph, weight_fn: lambda x: 1)
+///
+///     to return a weight of 1 for all edges. Also::
+///
+///         minimum_spanning_edges(graph, weight_fn: float)
+///
+///     to cast the edge object as a float as the weight.
+/// :param float default_weight: If ``weight_fn`` isn't specified this optional
+///     float value will be used for the weight/cost of each edge.
+///
+/// :returns: The :math:`N - |c|` edges of the Minimum Spanning Tree (or Forest, if :math:`|c| > 1`)
+///     where :math:`N` is the number of nodes and :math:`|c|` is the number of connected components of the graph
+/// :rtype: WeightedEdgeList
+#[pyfunction(weight_fn = "None", default_weight = "1.0")]
+#[pyo3(text_signature = "(graph, weight_fn=None, default_weight=1.0)")]
+pub fn minimum_spanning_edges(
+    py: Python,
+    graph: &graph::PyGraph,
+    weight_fn: Option<PyObject>,
+    default_weight: f64,
+) -> PyResult<WeightedEdgeList> {
+    let mut subgraphs = UnionFind::<usize>::new(graph.graph.node_bound());
+
+    let mut edge_list: Vec<(f64, EdgeReference<PyObject>)> =
+        Vec::with_capacity(graph.graph.edge_count());
+    for edge in graph.edge_references() {
+        let weight =
+            weight_callable(py, &weight_fn, edge.weight(), default_weight)?;
+        if weight.is_nan() {
+            return Err(PyValueError::new_err("NaN found as an edge weight"));
+        }
+        edge_list.push((weight, edge));
+    }
+
+    edge_list.par_sort_unstable_by(|a, b| {
+        let weight_a = a.0;
+        let weight_b = b.0;
+        weight_a.partial_cmp(&weight_b).unwrap_or(Ordering::Less)
+    });
+
+    let mut answer: Vec<(usize, usize, PyObject)> = Vec::new();
+    for float_edge_pair in edge_list.iter() {
+        let edge = float_edge_pair.1;
+        let u = edge.source().index();
+        let v = edge.target().index();
+        if subgraphs.union(u, v) {
+            let w = edge.weight().clone_ref(py);
+            answer.push((u, v, w));
+        }
+    }
+
+    Ok(WeightedEdgeList { edges: answer })
+}
+
+/// Find the minimum spanning tree or forest of a graph
+/// using Kruskal's algorithm.
+///
+/// :param PyGraph graph: Undirected graph
+/// :param weight_fn: A callable object (function, lambda, etc) which
+///     will be passed the edge object and expected to return a ``float``. This
+///     tells retworkx/rust how to extract a numerical weight as a ``float``
+///     for edge object. Some simple examples are::
+///
+///         minimum_spanning_tree(graph, weight_fn: lambda x: 1)
+///
+///     to return a weight of 1 for all edges. Also::
+///
+///         minimum_spanning_tree(graph, weight_fn: float)
+///
+///     to cast the edge object as a float as the weight.
+/// :param float default_weight: If ``weight_fn`` isn't specified this optional
+///     float value will be used for the weight/cost of each edge.
+///
+/// :returns: A Minimum Spanning Tree (or Forest, if the graph is not connected).
+///
+/// :rtype: PyGraph
+///
+/// .. note::
+///
+///     The new graph will keep the same node indexes, but edge indexes might differ.
+#[pyfunction(weight_fn = "None", default_weight = "1.0")]
+#[pyo3(text_signature = "(graph, weight_fn=None, default_weight=1.0)")]
+pub fn minimum_spanning_tree(
+    py: Python,
+    graph: &graph::PyGraph,
+    weight_fn: Option<PyObject>,
+    default_weight: f64,
+) -> PyResult<graph::PyGraph> {
+    let mut spanning_tree = (*graph).clone();
+    spanning_tree.graph.clear_edges();
+
+    for edge in minimum_spanning_edges(py, graph, weight_fn, default_weight)?
+        .edges
+        .iter()
+    {
+        spanning_tree.add_edge(edge.0, edge.1, edge.2.clone_ref(py))?;
+    }
+
+    Ok(spanning_tree)
+}
